requirements.txt
altair
certifi
Flask
Fortuna
gunicorn
Jinja2
joblib
jupyter
numpy
MonsterLab
pandas
pymongo[srv]
python-dotenv
scikit-learn
scipy
psycopg2
